[ taha.darendeli ][ Completed CatalougeView UI, Added tap action on collectionViewCell, Fixed naming issue on CatalougeView ]
