+++ Clothes Store/Controllers/CatalougeViewHostingViewController.swift
@@ -1,5 +1,5 @@
 //
-//  CatalogueViewHostingViewController.swift
+//  CatalougeViewHostingViewController.swift
 //  Clothes Store
 //
 //  Created by Mentor on 29.05.2021.
@@ -10,9 +10,9 @@ import Foundation
 import UIKit
 import SwiftUI
 
-final class CatalogueViewHostingViewController: UIHostingController<CatalogueView> {
+final class CatalougeViewHostingViewController: UIHostingController<CatalougeView> {
     required init?(coder aDecoder: NSCoder){
-        super.init(coder: aDecoder, rootView: CatalogueView())
+        super.init(coder: aDecoder, rootView: CatalougeView())
         
     }
 }
